song-downloader/requirements.txt
# song-downloader/requirements.txt
pika==1.3.2
yt-dlp==2024.10.22
grpcio==1.60.0
grpcio-tools==1.60.0
